# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "Constituent": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "relationType": {
            "title": "Relationtype",
            "type": "string"
          },
          "weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          }
        },
        "required": [
          "code",
          "name",
          "relationType",
          "weight"
        ],
        "title": "Constituent",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_TopicSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TopicSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[TopicSummary]",
        "type": "object"
      },
...
      "TopicConstituentRead": {
        "properties": {
          "changePct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changepct"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "factHash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facthash"
          },
          "factState": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Factstate"
          },
          "freshness": {
            "title": "Freshness",
            "type": "string"
          },
          "instrumentId": {
            "title": "Instrumentid",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "observationDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observationdate"
          },
          "observedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observedat"
          },
          "observedClassification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observedclassification"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "relationWeight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationweight"
          },
          "relativeTopicState": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relativetopicstate"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "technicalState": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technicalstate"
          },
          "updateMode": {
            "title": "Updatemode",
            "type": "string"
          }
        },
        "required": [
          "instrumentId",
          "symbol",
          "code",
          "name",
          "role",
          "relationWeight",
          "price",
          "changePct",
          "observedAt",
          "updateMode",
          "freshness",
          "technicalState",
          "relativeTopicState"
        ],
        "title": "TopicConstituentRead",
        "type": "object"
      },
...
      "TopicReadModel": {
        "properties": {
          "constituentCount": {
            "title": "Constituentcount",
            "type": "integer"
          },
          "constituents": {
            "items": {
              "$ref": "#/components/schemas/TopicConstituentRead"
            },
            "title": "Constituents",
            "type": "array"
          },
          "coveragePct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coveragepct"
          },
          "dataDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datadate"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "groupName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupname"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/TopicLifecycleRead"
          },
          "lineage": {
            "additionalProperties": true,
            "title": "Lineage",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "publication": {
            "additionalProperties": true,
            "title": "Publication",
            "type": "object"
          },
          "quality": {
            "additionalProperties": true,
            "title": "Quality",
            "type": "object"
          },
          "readableState": {
            "title": "Readablestate",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "items": {
              "$ref": "#/components/schemas/TopicStatusRead"
            },
            "title": "Status",
            "type": "array"
          },
          "strengthState": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strengthstate"
          },
          "topicId": {
            "title": "Topicid",
            "type": "string"
          },
          "topicType": {
            "title": "Topictype",
            "type": "string"
          }
        },
        "required": [
          "topicId",
          "slug",
          "name",
          "groupName",
          "topicType",
          "enabled",
          "dataDate",
          "score",
          "grade",
          "direction",
          "strengthState",
          "readableState",
          "coveragePct",
          "constituentCount",
          "lifecycle"
        ],
        "title": "TopicReadModel",
        "type": "object"
      },
      "TopicReadModelPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TopicReadModel"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "query"
        ],
        "title": "TopicReadModelPage",
        "type": "object"
      },
...
      "TopicResponse": {
        "properties": {
          "constituentCount": {
            "title": "Constituentcount",
            "type": "integer"
          },
          "constituents": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Constituent"
            },
            "title": "Constituents",
            "type": "array"
          },
          "coveragePct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coveragepct"
          },
          "dataDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datadate"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "groupName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupname"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "strengthState": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strengthstate"
          },
          "topicType": {
            "title": "Topictype",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "groupName",
          "topicType",
          "enabled",
          "dataDate",
          "score",
          "grade",
          "strengthState",
          "coveragePct",
          "constituentCount"
        ],
        "title": "TopicResponse",
        "type": "object"
      },
...
      "TopicSnapshotPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TopicSnapshotResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "query"
        ],
        "title": "TopicSnapshotPage",
        "type": "object"
      },
...
      "TopicStatusRead": {
        "properties": {
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "required": [
          "key",
          "state"
        ],
        "title": "TopicStatusRead",
        "type": "object"
      },
      "TopicSummary": {
        "properties": {
          "constituentCount": {
            "title": "Constituentcount",
            "type": "integer"
          },
          "coveragePct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coveragepct"
          },
          "dataDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datadate"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "groupName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupname"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "strengthState": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strengthstate"
          },
          "topicType": {
            "title": "Topictype",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "groupName",
          "topicType",
          "enabled",
          "dataDate",
          "score",
          "grade",
          "strengthState",
          "coveragePct",
          "constituentCount"
        ],
        "title": "TopicSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only API backed by a rebuildable PostgreSQL read model. The public deployment contains synthetic data only.",
    "title": "TopicPilot Enterprise Read API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/topics": {
      "get": {
        "operationId": "topics_api_v1_admin_topics_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Topics Api V1 Admin Topics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Topics",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/topics/{topic_id}": {
      "get": {
        "operationId": "topic_detail_api_v1_admin_topics__topic_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "topic_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Topic Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Topic Detail Api V1 Admin Topics  Topic Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Topic Detail",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/topics": {
      "get": {
        "operationId": "topics_api_v1_topics_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TopicSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Topics",
        "tags": [
          "topics"
        ]
      }
    },
    "/api/v1/topics/{slug}": {
      "get": {
        "operationId": "topic_api_v1_topics__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Topic",
        "tags": [
          "topics"
        ]
      }
    },
...
    "/api/v2/topic-snapshots": {
      "get": {
        "operationId": "list_topic_snapshots_api_v2_topic_snapshots_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "latest",
            "required": false,
            "schema": {
              "default": true,
              "title": "Latest",
              "type": "boolean"
            }
          },
          {
            "description": "Topic slug",
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Topic slug",
              "title": "Topic"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicSnapshotPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read V2 topic snapshots by date or history",
        "tags": [
          "topic-snapshots"
        ]
      }
    },
    "/api/v2/topics": {
      "get": {
        "operationId": "topics_api_v2_topics_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicReadModelPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read the formal topic read model",
        "tags": [
          "production-read-model"
        ]
      }
    },
    "/api/v2/topics/{slug}": {
      "get": {
        "operationId": "topic_api_v2_topics__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicReadModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one formal topic",
        "tags": [
          "production-read-model"
        ]
      }
    },
...
  }
}